LayoutTests/platform/mac-sonoma-wk1/imported/w3c/web-platform-tests/media-source/mediasource-seek-during-pending-seek-expected.txt


FAIL Test seeking to a new location before transitioning beyond HAVE_METADATA. assert_equals: Event types match. expected "seeked" but got "playing"
PASS Test seeking to a new location during a pending seek.
